Sprint 1: Move an object using the arrow keys
We add interaction by creating an event listener. This listens for keypress events and stores which key has been pressed.
We have created a game loop which is called every 100ms.
The game loop prints a square to the grid at the current player location.
The location can be changed using the arrow keys.
When the player location goes outside the grid it is wrapped around.